# perf: optimize process.nextTick with object pooling and batch processing