--- a/lib/internal/fixed_queue.js
+++ b/lib/internal/fixed_queue.js
@@ -9,6 +9,11 @@ const {
 const kSize = 2048;
 const kMask = kSize - 1;
 
+// Buffer pool to reduce allocations - optimization for nextTick performance
+const kMaxPoolSize = 16; // Keep reasonable number of buffers in pool
+let bufferPool = [];
+let bufferPoolSize = 0;
+
 // The FixedQueue is implemented as a singly-linked list of fixed-size
 // circular buffers. It looks something like this:
 //
@@ -57,6 +62,26 @@ const kMask = kSize - 1;
 // `top + 1 === bottom` it's full. This wastes a single space of storage
 // but allows much quicker checks.
 
+// Buffer pool functions for optimization
+function getPooledBuffer() {
+  if (bufferPoolSize > 0) {
+    const buffer = bufferPool[--bufferPoolSize];
+    // Reset buffer state for reuse
+    buffer.bottom = 0;
+    buffer.top = 0;
+    buffer.next = null;
+    ArrayPrototypeFill(buffer.list, undefined);
+    return buffer;
+  }
+  return new FixedCircularBuffer();
+}
+
+function releaseBuffer(buffer) {
+  if (bufferPoolSize < kMaxPoolSize) {
+    bufferPool[bufferPoolSize++] = buffer;
+  }
+}
+
 class FixedCircularBuffer {
   constructor() {
     this.bottom = 0;
@@ -86,11 +111,27 @@ class FixedCircularBuffer {
     this.bottom = (this.bottom + 1) & kMask;
     return nextItem;
   }
+
+  // Optimized batch shift for better cache locality
+  shiftBatch(destination, maxCount) {
+    let count = 0;
+    while (count < maxCount && !this.isEmpty()) {
+      const item = this.list[this.bottom];
+      if (item === undefined) break;
+      
+      destination[count++] = item;
+      this.list[this.bottom] = undefined;
+      this.bottom = (this.bottom + 1) & kMask;
+    }
+    return count;
+  }
 }
 
 module.exports = class FixedQueue {
   constructor() {
     this.head = this.tail = new FixedCircularBuffer();
+    // Batch processing buffer for optimization
+    this.batchBuffer = new Array(64);
   }
 
   isEmpty() {
@@ -99,9 +140,8 @@ module.exports = class FixedQueue {
 
   push(data) {
     if (this.head.isFull()) {
-      // Head is full: Creates a new queue, sets the old queue's `.next` to it,
-      // and sets it as the new main queue.
-      this.head = this.head.next = new FixedCircularBuffer();
+      // Head is full: Use pooled buffer instead of always creating new
+      this.head = this.head.next = getPooledBuffer();
     }
     this.head.push(data);
   }
@@ -110,10 +150,40 @@ module.exports = class FixedQueue {
     const tail = this.tail;
     const next = tail.shift();
     if (tail.isEmpty() && tail.next !== null) {
-      // If there is another queue, it forms the new tail.
+      // Return old tail to pool for reuse
+      releaseBuffer(this.tail);
       this.tail = tail.next;
       tail.next = null;
     }
     return next;
+  }
+
+  // Optimized batch processing method
+  processBatch(processor, maxBatch = 64) {
+    let processed = 0;
+    
+    while (processed < maxBatch && !this.isEmpty()) {
+      const tail = this.tail;
+      const batchCount = tail.shiftBatch(this.batchBuffer, maxBatch - processed);
+      
+      if (batchCount === 0) break;
+      
+      // Process batch with better cache locality
+      for (let i = 0; i < batchCount; i++) {
+        processor(this.batchBuffer[i]);
+        this.batchBuffer[i] = undefined; // Clear reference
+      }
+      
+      processed += batchCount;
+      
+      // Check if we need to move to next buffer
+      if (tail.isEmpty() && tail.next !== null) {
+        releaseBuffer(this.tail);
+        this.tail = tail.next;
+        tail.next = null;
+      }
+    }
+    
+    return processed;
   }
 };