--- a/test/parallel/test-next-tick-optimization-coverage.js
+++ b/test/parallel/test-next-tick-optimization-coverage.js
@@ -1,0 +1,242 @@
+'use strict';
+
+// Test coverage for nextTick performance optimizations
+// This test ensures the object pooling, args pooling, and batch processing
+// code paths are properly covered in the optimization commit.
+
+const common = require('../common');
+const assert = require('assert');
+
+let testCount = 0;
+let completedTests = 0;
+
+function completeTest(testName) {
+  completedTests++;
+  console.log(`✓ ${testName} completed (${completedTests}/${testCount})`);
+}
+
+// Test 1: Object pooling exhaustion and reuse
+// Exceeds kTickObjectPoolSize (512) to test both pool hit/miss paths
+testCount++;
+{
+  let completedCount = 0;
+  const totalCalls = 600; // Exceeds pool size of 512
+  
+  for (let i = 0; i < totalCalls; i++) {
+    process.nextTick(() => {
+      completedCount++;
+      if (completedCount === totalCalls) {
+        assert.strictEqual(completedCount, totalCalls);
+        completeTest('Object pooling exhaustion test');
+      }
+    });
+  }
+}
+
+// Test 2: Args pooling with different argument counts (0-8+ args)
+// Tests all switch statement cases and fallback paths
+testCount++;
+{
+  let completedArgTests = 0;
+  const expectedArgTests = 10; // 0-8 args + >8 args case
+  
+  // 0 args - tests undefined args path
+  process.nextTick(() => {
+    completedArgTests++;
+    if (completedArgTests === expectedArgTests) {
+      completeTest('Args pooling with different argument counts');
+    }
+  });
+  
+  // 1 arg
+  process.nextTick((a) => {
+    assert.strictEqual(a, 1);
+    completedArgTests++;
+    if (completedArgTests === expectedArgTests) {
+      completeTest('Args pooling with different argument counts');
+    }
+  }, 1);
+  
+  // 2 args
+  process.nextTick((a, b) => {
+    assert.strictEqual(a, 1);
+    assert.strictEqual(b, 2);
+    completedArgTests++;
+    if (completedArgTests === expectedArgTests) {
+      completeTest('Args pooling with different argument counts');
+    }
+  }, 1, 2);
+  
+  // 3 args
+  process.nextTick((a, b, c) => {
+    assert.strictEqual(a, 1);
+    assert.strictEqual(b, 2);
+    assert.strictEqual(c, 3);
+    completedArgTests++;
+    if (completedArgTests === expectedArgTests) {
+      completeTest('Args pooling with different argument counts');
+    }
+  }, 1, 2, 3);
+  
+  // 4 args
+  process.nextTick((a, b, c, d) => {
+    assert.strictEqual(a, 1);
+    assert.strictEqual(b, 2);
+    assert.strictEqual(c, 3);
+    assert.strictEqual(d, 4);
+    completedArgTests++;
+    if (completedArgTests === expectedArgTests) {
+      completeTest('Args pooling with different argument counts');
+    }
+  }, 1, 2, 3, 4);
+  
+  // 5 args
+  process.nextTick((a, b, c, d, e) => {
+    assert.strictEqual(a, 1);
+    assert.strictEqual(b, 2);
+    assert.strictEqual(c, 3);
+    assert.strictEqual(d, 4);
+    assert.strictEqual(e, 5);
+    completedArgTests++;
+    if (completedArgTests === expectedArgTests) {
+      completeTest('Args pooling with different argument counts');
+    }
+  }, 1, 2, 3, 4, 5);
+  
+  // 6 args
+  process.nextTick((a, b, c, d, e, f) => {
+    assert.strictEqual(a, 1);
+    assert.strictEqual(b, 2);
+    assert.strictEqual(c, 3);
+    assert.strictEqual(d, 4);
+    assert.strictEqual(e, 5);
+    assert.strictEqual(f, 6);
+    completedArgTests++;
+    if (completedArgTests === expectedArgTests) {
+      completeTest('Args pooling with different argument counts');
+    }
+  }, 1, 2, 3, 4, 5, 6);
+  
+  // 7 args
+  process.nextTick((a, b, c, d, e, f, g) => {
+    assert.strictEqual(a, 1);
+    assert.strictEqual(b, 2);
+    assert.strictEqual(c, 3);
+    assert.strictEqual(d, 4);
+    assert.strictEqual(e, 5);
+    assert.strictEqual(f, 6);
+    assert.strictEqual(g, 7);
+    completedArgTests++;
+    if (completedArgTests === expectedArgTests) {
+      completeTest('Args pooling with different argument counts');
+    }
+  }, 1, 2, 3, 4, 5, 6, 7);
+  
+  // 8 args
+  process.nextTick((a, b, c, d, e, f, g, h) => {
+    assert.strictEqual(a, 1);
+    assert.strictEqual(b, 2);
+    assert.strictEqual(c, 3);
+    assert.strictEqual(d, 4);
+    assert.strictEqual(e, 5);
+    assert.strictEqual(f, 6);
+    assert.strictEqual(g, 7);
+    assert.strictEqual(h, 8);
+    completedArgTests++;
+    if (completedArgTests === expectedArgTests) {
+      completeTest('Args pooling with different argument counts');
+    }
+  }, 1, 2, 3, 4, 5, 6, 7, 8);
+  
+  // >8 args - tests fallback path with spread operator
+  process.nextTick((...args) => {
+    assert.strictEqual(args.length, 10);
+    for (let i = 0; i < 10; i++) {
+      assert.strictEqual(args[i], i + 1);
+    }
+    completedArgTests++;
+    if (completedArgTests === expectedArgTests) {
+      completeTest('Args pooling with different argument counts');
+    }
+  }, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10);
+}
+
+// Test 3: Args pool exhaustion 
+// Exceeds kArgsPoolSize (256) to test pool reuse
+testCount++;
+{
+  let poolTestCompleted = 0;
+  const poolTestCount = 300; // Exceeds args pool size of 256
+  
+  for (let i = 0; i < poolTestCount; i++) {
+    process.nextTick((a, b, c) => {
+      assert.strictEqual(a, i);
+      assert.strictEqual(b, i + 1);
+      assert.strictEqual(c, i + 2);
+      poolTestCompleted++;
+      
+      if (poolTestCompleted === poolTestCount) {
+        assert.strictEqual(poolTestCompleted, poolTestCount);
+        completeTest('Args pool exhaustion test');
+      }
+    }, i, i + 1, i + 2);
+  }
+}
+
+// Test 4: High-volume batch processing
+// Tests batch processing paths and queue buffer management
+testCount++;
+{
+  let batchTestCompleted = 0;
+  const batchTestCount = 1000; // High volume to trigger batch processing
+  
+  for (let i = 0; i < batchTestCount; i++) {
+    process.nextTick(() => {
+      batchTestCompleted++;
+      
+      if (batchTestCompleted === batchTestCount) {
+        assert.strictEqual(batchTestCompleted, batchTestCount);
+        completeTest('High-volume batch processing test');
+      }
+    });
+  }
+}
+
+// Test 5: Mixed patterns to stress all optimization systems
+testCount++;
+{
+  let mixedTestCompleted = 0;
+  const mixedTestCount = 200;
+  
+  for (let i = 0; i < mixedTestCount; i++) {
+    // Alternate between different argument counts and patterns
+    const argCount = i % 9; // 0-8 args
+    
+    if (argCount === 0) {
+      process.nextTick(() => {
+        mixedTestCompleted++;
+        if (mixedTestCompleted === mixedTestCount) {
+          completeTest('Mixed patterns stress test');
+        }
+      });
+    } else if (argCount <= 8) {
+      const args = new Array(argCount).fill(0).map((_, idx) => i + idx);
+      process.nextTick((...receivedArgs) => {
+        assert.strictEqual(receivedArgs.length, argCount);
+        for (let j = 0; j < argCount; j++) {
+          assert.strictEqual(receivedArgs[j], i + j);
+        }
+        mixedTestCompleted++;
+        if (mixedTestCompleted === mixedTestCount) {
+          completeTest('Mixed patterns stress test');
+        }
+      }, ...args);
+    }
+  }
+}
+
+process.on('exit', () => {
+  assert.strictEqual(completedTests, testCount, 
+    `Expected ${testCount} tests to complete, but only ${completedTests} completed`);
+  console.log('✓ All nextTick optimization coverage tests completed successfully');
+});